orchestrator/requirements.txt
pydantic-ai[anthropic,openai,google,ollama]>=0.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic-settings>=2.7.0
httpx[http2]>=0.27.0
websockets>=14.0
mcp>=1.0.0
sqlalchemy>=2.0.25
geoalchemy2>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
asyncpg>=0.29.0
tenacity>=8.2.3
aiocache>=0.12.2
